==> 2_train_ILM.ipynb <==
The code for ILM is included in this repository, but the original and the setup instructions can be found [here](https://github.com/chrisdonahue/ilm). Before we can train the model, the compound dataset that we've created in the previous notebook needs be preprocessed: 

!cd ilm
!sh create_compound_dataset.sh

Then, we can run the training:

%%bash
DATASET=compound_dataset
TRAIN_DIR=../Models/ILM
EXAMPLES_DIR=data/char_masks/${DATASET}
python train_ilm.py \
	experiment_${DATASET} \
	${TRAIN_DIR} \
	${EXAMPLES_DIR} \
	--train_examples_tag train \
	--eval_examples_tag valid \
	--train_num_epochs 4 \
	--eval_max_num_examples 512
